:recycle: [Refactor]: FSD 구조 적용 - 7. widgets 레이어로 분리하기
