+++ apps/client/src/app/layouts/Layout.tsx
@@ -1,6 +1,7 @@
 import { Outlet } from 'react-router-dom';
 import { Toaster } from '@/shared/ui/shadcn/toaster';
-import { Header, FloatingButton } from '@/shared/ui';
+import { Header } from '@/widgets';
+import { FloatingButton } from '@/shared/ui';
 import { Providers } from '../providers';
 
 export function Layout() {
