+++ apps/client/src/features/editProfile/ui/EditUserInfo.tsx
@@ -1,7 +1,7 @@
 import { useForm } from 'react-hook-form';
 import { useState } from 'react';
 import { Avatar, AvatarFallback, AvatarImage } from '@/shared/ui/shadcn/avatar';
-import { UserData } from '@/entities/user';
+import { UserData } from '@/pages/Profile';
 import { Field } from '@/shared/types/sharedTypes';
 import { Button } from '@/shared/ui/shadcn/button';
 import { axiosInstance } from '@/shared/api';
