--- apps/client/src/pages/Home/ui/index.ts
@@ -1,5 +1,0 @@
-export { Banner } from './Banner';
-export { FieldFilter } from './FieldFilter';
-export { LiveCard } from './LiveCard';
-export { LiveList } from './LiveList';
-export { Search } from './Search';
