+++ apps/client/src/pages/Profile/index.ts
@@ -1,1 +1,2 @@
 export { ProfilePage } from './ProfilePage';
+export type { UserData } from './model';
