+++ apps/client/src/widgets/Banner/Banner.tsx
