+++ apps/client/src/widgets/Banner/Bookmark.tsx
