+++ apps/client/src/widgets/Banner/index.ts
