+++ apps/client/src/widgets/Banner/types.ts
