+++ apps/client/src/widgets/Header/WelcomeCharacter.tsx
