+++ apps/client/src/widgets/LiveList/LiveCard.tsx
