+++ apps/client/src/widgets/LiveList/index.ts
@@ -1,0 +1,1 @@
+export { LiveList } from './LiveList';
